+++ Package.swift
@@ -14,8 +14,8 @@ let package = Package(
         .library(name: "FeatherRelationalDatabaseDriverPostgres", targets: ["FeatherRelationalDatabaseDriverPostgres"]),
     ],
     dependencies: [
-        .package(url: "https://github.com/feather-framework/feather-relational-database", .upToNextMinor(from: "0.1.1")),
-        .package(url: "https://github.com/vapor/postgres-kit", from: "2.12.0"),
+        .package(url: "https://github.com/feather-framework/feather-relational-database", .upToNextMinor(from: "0.2.0")),
+        .package(url: "https://github.com/vapor/postgres-kit", from: "2.0.0"),
     ],
     targets: [
         .target(
